⚡ Prepare asyncness
Creating a bunch of futures that when completed get the buffers copied
to the screen buffer updating the screen as it gets done.

It's a bit overkill to create a future per row but it can be changed
later.

Moved hsv_to_rgb to utils. Don't even think it will be needed later.