📖 Add logging and term writes
println works for a while. Was time to set up something better. Worth
to not that there is a big difference between logging and writing to
the terminal which is why both slog and console was dragged in. Might
seem similar but purpose is not the same.

Most of the time the log is interesting during runtime but user
messages does not belong in the log.