✨ Add scene controller
Wanted to be able to for example move around freely in the scene but
also have something that would for example follow key frames and
render a gif.

Abstracted it with a scene controller. You can hook up keybinds and
other thigns for it as well. Right now there is only the interactive
scene controller which keep the behaviours previous to this change.

Now I could possibly switch it out with something that uses key frames
to render several images to create for example a gif.

List of other Misc changes:
- Add configuration setting for scene controller (`scene_controller`)
- Add configuration setting for renderer
- Add configuration setting for preview renderer (`preview_renderer`)
- Add clone to Config.
- Add from implementation for Renderer to be created from the config
  object.
- Add cancel event to image action. An action could be blocking when
  the application wants to exit. Actions can now listen to the cancel
  event to exit early and not block.
- Fixed bug where WaitForSignal action would block after application
  tries to exit.
- Add method to KeyInputs to be able to take a list of callbacks instead of
  manually registering every callback one at the time.