✨ Add BVH solution to scene
Had to refactor quite a bit

- Use message passing to update scene between threads.

While implementing the aabb hirearchy I noticed a need to change data
that should be reflected between threads.

You could just add a lock but it's pretty annoying using locks
everywhere. Added DataBus which implements MPMC. It's a bit overkill
since most cases should be covered by SPMC. Just wanted to have
something that would work with everything.

This means that the update thread can update whatever implements this
pattern and the changes will be reflected in the scene.

There is one caveat and that is that it takes a frame for the changes
to be propagated through the bus. It mostly matters for the real time
part but that part isn't super important to the resulting product.

A renderer is usually isolated to be able to produce an image. This
project is a harness around a part that can produce images.  The
harness (real-time part) is just less important. The latecy is not
that noteable either.