🎨 Change buffers to be colors for longer.
This change just prepares the impl of SVGF.
It postones the conversion of the Color to the color the screen buffer
expects which makes sense.

It also ruins some other things such as real time updates of chunks of
the screen when rendering the scene.

Don't think it's a big deal since SVGF is supposed to give real time
updates increasing the quality of the screen real time.

This also introduces more copying but it makes sense since certain
parts of the program can't agree on a format of data (because it
doesn't make any sense to do so).

Renderers now produce a float3. The screenbuffer will have to use that
to convert the data to something that makes sense to the screen buffer.